Move bake command into own file
